refactor(common): extract translationResolver from lookupByPolicy
Split the write-side orchestration (locale-direct lookup, policy
fallback, dedup, dispatch) out of lookupByPolicy.js into a dedicated
translationResolver.js. lookupByPolicy.js now holds only the DB
lookup strategy; translationResolver holds the orchestration.

buildResolver(db, translations, options) replaces the former
detectCommonData + buildPolicy + createLookupParams three-step setup
and has no side effects — common-project info is captured into the
resolver at build time instead of mutated onto the FileType instance.

Plugins (c, cpp, dart, javascript) call buildResolver + resolveTranslation.
