+++ .changeset/puny-rules-matter.md
@@ -1,0 +1,11 @@
+---
+"ilib-loctool-webos-dist": patch
+---
+
+ilib-loctool-webos-javascript
+ilib-loctool-webos-common
+ilib-loctool-webos-dart
+ilib-loctool-webos-cpp
+ilib-loctool-webos-c
+- Centralize common DB lookup logic from FileType.write()
+Replace per-plugin ResourceString.hashKey manual lookups with a shared lookupByPolicy engine in ilib-loctool-webos-common
