+++ packages/ilib-loctool-webos-common/index.js
@@ -19,5 +19,6 @@
 
 module.exports = {
     utils: require("./utils.js"),
-    lookupByPolicy: require("./lookupByPolicy.js")
+    lookupByPolicy: require("./lookupByPolicy.js"),
+    translationResolver: require("./translationResolver.js")
 };
