+++ packages/ilib-loctool-webos-common/test/lookupByPolicy.test.js
@@ -17,8 +17,7 @@
  * limitations under the License.
  */
 
-var { buildPolicy, lookupByPolicy, createLookupParams, detectCommonData, writeTranslatedResource } = require("../lookupByPolicy.js");
-var TranslationSet = require("loctool/lib/TranslationSet.js");
+var { buildPolicy, lookupByPolicy } = require("../lookupByPolicy.js");
 var ResourceString = require("loctool/lib/ResourceString.js");
 var Utils = require("loctool/lib/utils.js");
 
@@ -127,592 +126,5 @@ describe("lookupByPolicy", function() {
 
 });
 
-// ── createLookupParams ────────────────────────────────────────────────────────
 
-describe("createLookupParams", function() {
-    test("builds a params object carrying the shared context and per-lookup fields", function() {
-        var db = makeDb({});
-        var policy = buildPolicy();
-        var fileType = {
-            commonPrjName: "common",
-            commonPrjType: "x-json"
-        };
-        var resource = makeResource("myapp", "hello");
 
-        var make = createLookupParams(fileType, db, policy);
-        var params = make(resource, "ko-KR");
-
-        expect(params).toEqual({
-            db: db,
-            resource: resource,
-            locale: "ko-KR",
-            policy: policy,
-            commonPrjName: "common",
-            commonPrjType: "x-json"
-        });
-    });
-
-    test("reads commonPrjName/Type lazily from fileType on each call", function() {
-        var db = makeDb({});
-        var policy = buildPolicy();
-        var fileType = {};
-
-        var make = createLookupParams(fileType, db, policy);
-        var before = make(makeResource("myapp", "a"), "ko-KR");
-        expect(before.commonPrjName).toBeUndefined();
-        expect(before.commonPrjType).toBeUndefined();
-
-        // common data populated later during write()
-        fileType.commonPrjName = "common";
-        fileType.commonPrjType = "x-json";
-
-        var after = make(makeResource("myapp", "b"), "ja-JP");
-        expect(after.commonPrjName).toBe("common");
-        expect(after.commonPrjType).toBe("x-json");
-        expect(after.locale).toBe("ja-JP");
-    });
-
-    test("produced params drive lookupByPolicy to a common-step hit", function(done) {
-        var ResourceString = require("loctool/lib/ResourceString.js");
-        var commonKey = ResourceString.hashKey("common", "ko-KR", "hello", "x-json", undefined);
-        var fakeCommon = { target: "공통 안녕" };
-        var db = makeDb({ [commonKey]: fakeCommon });
-        var fileType = {
-            commonPrjName: "common",
-            commonPrjType: "x-json"
-        };
-
-        var make = createLookupParams(fileType, db, buildPolicy());
-        lookupByPolicy(make(makeResource("myapp", "hello"), "ko-KR"), function(result) {
-            expect(result).toBe(fakeCommon);
-            done();
-        });
-    });
-});
-
-// ── detectCommonData ──────────────────────────────────────────────────────────
-
-describe("detectCommonData", function() {
-    function makeCommonTs() {
-        var ts = new TranslationSet();
-        ts.add(new ResourceString({
-            project: "common",
-            sourceLocale: "en-US",
-            targetLocale: "ko-KR",
-            key: "hello",
-            source: "hello",
-            target: "안녕",
-            datatype: "x-json"
-        }));
-        return ts;
-    }
-
-    test("translations undefined — no-op, commonPrjName not set", function() {
-        var ft = {};
-        detectCommonData(ft, undefined);
-        expect(ft.commonPrjName).toBeUndefined();
-    });
-
-    test("no common project in translations — commonPrjName not set", function() {
-        var ts = new TranslationSet();
-        ts.add(new ResourceString({
-            project: "myapp",
-            sourceLocale: "en-US",
-            targetLocale: "ko-KR",
-            key: "hi",
-            source: "hi",
-            target: "안녕"
-        }));
-        var ft = {};
-        detectCommonData(ft, ts);
-        expect(ft.commonPrjName).toBeUndefined();
-    });
-
-    test("common project present — sets commonPrjName and commonPrjType", function() {
-        var ts = makeCommonTs();
-        var ft = {};
-        detectCommonData(ft, ts);
-        expect(ft.commonPrjName).toBe("common");
-        expect(ft.commonPrjType).toBe("x-json");
-    });
-
-    test("common project present but getBy returns empty — name/type not set", function() {
-        var ts = new TranslationSet();
-        ts.getProjects = function() { return ["common"]; };
-        ts.getBy = function() { return []; };
-        var ft = {};
-        detectCommonData(ft, ts);
-        expect(ft.commonPrjName).toBeUndefined();
-        expect(ft.commonPrjType).toBeUndefined();
-    });
-});
-
-// ── writeTranslatedResource ───────────────────────────────────────────────────
-
-describe("writeTranslatedResource", function() {
-    function makeTestResource(key, source) {
-        return {
-            reskey: "rk-" + key,
-            getProject: function() { return "myapp"; },
-            getDataType: function() { return "cpp"; },
-            getPath: function() { return "src/main.cpp"; },
-            getContext: function() { return "main"; },
-            getKey: function() { return key; },
-            getSource: function() { return source; },
-            getFlavor: function() { return undefined; },
-            cleanHashKeyForTranslation: function(locale) { return key + "::" + locale; },
-            clone: function() {
-                var copy = makeTestResource(key, source);
-                copy.setTargetLocale = function(locale) { this._targetLocale = locale; };
-                copy.setTarget = function(target) { this.target = target; };
-                copy.setState = function(state) { this.state = state; };
-                copy.setComment = function(comment) { this.comment = comment; };
-                return copy;
-            }
-        };
-    }
-
-    function makeTranslated(target, source, key) {
-        var make = function(t, s, k) {
-            var obj = {
-                reskey: "tr-" + (k || "key"),
-                getSource: function() { return s || "src"; },
-                getKey: function() { return k || "key"; },
-                getFlavor: function() { return undefined; },
-                target: t,
-                metadata: { target: t },
-                setTargetLocale: function(locale) { this._targetLocale = locale; },
-                setTarget: function(value) { this.target = value; }
-            };
-            obj.clone = function() {
-                var copy = make(this.target, this.getSource(), this.getKey());
-                copy.reskey = this.reskey;
-                return copy;
-            };
-            return obj;
-        };
-
-        return make(target, source, key);
-    }
-
-    function makeApi() {
-        return {
-            utils: {
-                cleanString: function(s) { return s; }
-            }
-        };
-    }
-
-    function makeCollector() {
-        var map = {};
-        return {
-            getResourceFile: function(locale, resPath) {
-                var key = resPath || locale;
-                if (!map[key]) {
-                    map[key] = {
-                        pathName: key + ".json",
-                        resources: [],
-                        addResource: function(r) { this.resources.push(r); }
-                    };
-                }
-                return map[key];
-            },
-            _files: map
-        };
-    }
-
-    function makeNewres() {
-        var added = [];
-        return {
-            add: jest.fn(function(r) { added.push(r); }),
-            _added: added
-        };
-    }
-
-    // ── dedup mode tests ──────────────────────────────────────────────────────
-
-    test("dedupByBaseTranslation false writes even when target equals source", function(done) {
-        var res = makeTestResource("hello", "Hello");
-        var translated = makeTranslated("Hello", "Hello", "Hello");
-        var db = {
-            getResourceByCleanHashKey: jest.fn(function(key, cb) {
-                cb(null, key === "hello::fr-FR" ? translated : null);
-            })
-        };
-        var resFileType = makeCollector();
-        var newres = makeNewres();
-        var makeLookupParams = function(resource, locale) {
-            return { db: db, resource: resource, locale: locale, policy: [] };
-        };
-
-        writeTranslatedResource({
-            db: db,
-            resFileType: resFileType,
-            newres: newres,
-            res: res,
-            locale: "fr-FR",
-            dedupByBaseTranslation: false,
-            deviceType: undefined,
-            API: makeApi(),
-            makeLookupParams: makeLookupParams
-        }, function() {
-            expect(resFileType._files["fr-FR"].resources.length).toBe(1);
-            expect(newres.add).not.toHaveBeenCalled();
-            done();
-        });
-    });
-
-    test("dedupByBaseTranslation true + langDefault resolves base and skips same target", function(done) {
-        var res = makeTestResource("hello", "Hello");
-        var frTranslated = makeTranslated("Hola", "Hello", "Hello");
-        var baseTranslated = makeTranslated("Hola", "Hello", "Hello");
-        var locale = "es-CO";
-        var langDefaultLocale = Utils.getBaseLocale(locale);
-        var db = {
-            getResourceByCleanHashKey: jest.fn(function(key, cb) {
-                if (key === "hello::" + langDefaultLocale) {
-                    cb(null, baseTranslated);
-                    return;
-                }
-                if (key === "hello::" + locale) {
-                    cb(null, frTranslated);
-                    return;
-                }
-                cb(null, null);
-            })
-        };
-        var resFileType = makeCollector();
-        var newres = makeNewres();
-        var makeLookupParams = function(resource, locale) {
-            return { db: db, resource: resource, locale: locale, policy: [] };
-        };
-
-        writeTranslatedResource({
-            db: db,
-            resFileType: resFileType,
-            newres: newres,
-            res: res,
-            locale: locale,
-            dedupByBaseTranslation: true,
-            translationLocales: [langDefaultLocale, locale],
-            deviceType: undefined,
-            API: makeApi(),
-            makeLookupParams: makeLookupParams
-        }, function() {
-            expect(resFileType._files[locale]).toBeUndefined();
-            expect(newres.add).not.toHaveBeenCalled();
-            done();
-        });
-    });
-
-    test("explicit baseTranslation is used even when translationLocales is present", function(done) {
-        var res = makeTestResource("hello", "Hello");
-        var frTranslated = makeTranslated("Hola", "Hello", "Hello");
-        var baseTranslated = makeTranslated("Bonjour", "Hello", "Hello");
-        var locale = "es-CO";
-        var langDefaultLocale = Utils.getBaseLocale(locale);
-        var db = {
-            getResourceByCleanHashKey: jest.fn(function(key, cb) {
-                if (key === "hello::" + langDefaultLocale) {
-                    cb(null, baseTranslated);
-                    return;
-                }
-                if (key === "hello::" + locale) {
-                    cb(null, frTranslated);
-                    return;
-                }
-                cb(null, null);
-            })
-        };
-        var resFileType = makeCollector();
-        var newres = makeNewres();
-        var makeLookupParams = function(resource, locale) {
-            return { db: db, resource: resource, locale: locale, policy: [] };
-        };
-
-        writeTranslatedResource({
-            db: db,
-            resFileType: resFileType,
-            newres: newres,
-            res: res,
-            locale: locale,
-            dedupByBaseTranslation: true,
-            baseTranslation: "Hola",
-            translationLocales: [langDefaultLocale, locale],
-            deviceType: undefined,
-            API: makeApi(),
-            makeLookupParams: makeLookupParams
-        }, function() {
-            expect(db.getResourceByCleanHashKey).not.toHaveBeenCalledWith("hello::" + langDefaultLocale, expect.any(Function));
-            expect(resFileType._files[locale]).toBeUndefined();
-            done();
-        });
-    });
-
-    test("dedup uses source as fallback base when translationLocales is absent", function(done) {
-        var res = makeTestResource("hello", "Hello");
-        var translated = makeTranslated("Hello", "Hello", "Hello");
-        var db = {
-            getResourceByCleanHashKey: jest.fn(function(key, cb) {
-                cb(null, key === "hello::fr-FR" ? translated : null);
-            })
-        };
-        var resFileType = makeCollector();
-        var newres = makeNewres();
-        var makeLookupParams = function(resource, locale) {
-            return { db: db, resource: resource, locale: locale, policy: [] };
-        };
-
-        writeTranslatedResource({
-            db: db,
-            resFileType: resFileType,
-            newres: newres,
-            res: res,
-            locale: "fr-FR",
-            dedupByBaseTranslation: true,
-            deviceType: undefined,
-            API: makeApi(),
-            makeLookupParams: makeLookupParams
-        }, function() {
-            expect(resFileType._files["fr-FR"]).toBeUndefined();
-            done();
-        });
-    });
-
-    // ── direct miss → policy hit ──────────────────────────────────────────────
-
-    test("direct miss + policy hit + differs from base → addResource", function(done) {
-        var res = makeTestResource("hello", "Hello");
-        var commonTranslated = makeTranslated("안녕", "Hello", "hello");
-        var policy = buildPolicy();
-        var commonKey = ResourceString.hashKey("common", "ko-KR", "hello", "x-json", undefined);
-        var db = {
-            getResourceByCleanHashKey: jest.fn(function(key, cb) {
-                cb(null, key === commonKey ? commonTranslated : null);
-            })
-        };
-        var resFileType = makeCollector();
-        var newres = makeNewres();
-        var makeLookupParams = createLookupParams(
-            { commonPrjName: "common", commonPrjType: "x-json" },
-            db,
-            policy
-        );
-
-        writeTranslatedResource({
-            db: db,
-            resFileType: resFileType,
-            newres: newres,
-            res: res,
-            locale: "ko-KR",
-            dedupByBaseTranslation: true,
-            deviceType: undefined,
-            API: makeApi(),
-            makeLookupParams: makeLookupParams
-        }, function() {
-            expect(resFileType._files["ko-KR"].resources.length).toBe(1);
-            expect(newres.add).not.toHaveBeenCalled();
-            done();
-        });
-    });
-
-    test("direct miss + policy hit + same as base → addNewResource (skip)", function(done) {
-        var res = makeTestResource("hello", "Hello");
-        var commonTranslated = makeTranslated("Hello", "Hello", "hello");
-        var policy = buildPolicy();
-        var commonKey = ResourceString.hashKey("common", "ko-KR", "hello", "x-json", undefined);
-        var db = {
-            getResourceByCleanHashKey: jest.fn(function(key, cb) {
-                cb(null, key === commonKey ? commonTranslated : null);
-            })
-        };
-        var resFileType = makeCollector();
-        var newres = makeNewres();
-        var makeLookupParams = createLookupParams(
-            { commonPrjName: "common", commonPrjType: "x-json" },
-            db,
-            policy
-        );
-
-        writeTranslatedResource({
-            db: db,
-            resFileType: resFileType,
-            newres: newres,
-            res: res,
-            locale: "ko-KR",
-            dedupByBaseTranslation: true,
-            deviceType: undefined,
-            API: makeApi(),
-            makeLookupParams: makeLookupParams
-        }, function() {
-            expect(resFileType._files["ko-KR"]).toBeUndefined();
-            expect(newres.add).toHaveBeenCalledTimes(1);
-            done();
-        });
-    });
-
-    // ── direct miss + policy miss → customInherit fallback ────────────────────
-
-    test("direct miss + policy miss + customInherit direct hit → addResource", function(done) {
-        var res = makeTestResource("hello", "Hello");
-        var inheritTranslated = makeTranslated("Hola", "Hello", "hello");
-        var db = {
-            getResourceByCleanHashKey: jest.fn(function(key, cb) {
-                cb(null, key === "hello::es" ? inheritTranslated : null);
-            })
-        };
-        var resFileType = makeCollector();
-        var newres = makeNewres();
-        var makeLookupParams = function(resource, locale) {
-            return { db: db, resource: resource, locale: locale, policy: [] };
-        };
-
-        writeTranslatedResource({
-            db: db,
-            resFileType: resFileType,
-            newres: newres,
-            res: res,
-            locale: "es-CO",
-            customInheritLocale: "es",
-            dedupByBaseTranslation: true,
-            deviceType: undefined,
-            API: makeApi(),
-            makeLookupParams: makeLookupParams
-        }, function() {
-            expect(resFileType._files["es-CO"].resources.length).toBe(1);
-            expect(newres.add).not.toHaveBeenCalled();
-            done();
-        });
-    });
-
-    test("direct miss + policy miss + customInherit direct miss + inherit policy hit → addResource", function(done) {
-        var res = makeTestResource("hello", "Hello");
-        var policy = buildPolicy();
-        var inheritCommonKey = ResourceString.hashKey("common", "es", "hello", "x-json", undefined);
-        var inheritTranslated = makeTranslated("Hola", "Hello", "hello");
-        var db = {
-            getResourceByCleanHashKey: jest.fn(function(key, cb) {
-                cb(null, key === inheritCommonKey ? inheritTranslated : null);
-            })
-        };
-        var resFileType = makeCollector();
-        var newres = makeNewres();
-        var makeLookupParams = createLookupParams(
-            { commonPrjName: "common", commonPrjType: "x-json" },
-            db,
-            policy
-        );
-
-        writeTranslatedResource({
-            db: db,
-            resFileType: resFileType,
-            newres: newres,
-            res: res,
-            locale: "es-CO",
-            customInheritLocale: "es",
-            dedupByBaseTranslation: true,
-            deviceType: undefined,
-            API: makeApi(),
-            makeLookupParams: makeLookupParams
-        }, function() {
-            expect(resFileType._files["es-CO"].resources.length).toBe(1);
-            expect(newres.add).not.toHaveBeenCalled();
-            done();
-        });
-    });
-
-    test("all lookups miss → addNewResource", function(done) {
-        var res = makeTestResource("hello", "Hello");
-        var db = {
-            getResourceByCleanHashKey: jest.fn(function(key, cb) { cb(null, null); })
-        };
-        var resFileType = makeCollector();
-        var newres = makeNewres();
-        var makeLookupParams = function(resource, locale) {
-            return { db: db, resource: resource, locale: locale, policy: [] };
-        };
-
-        writeTranslatedResource({
-            db: db,
-            resFileType: resFileType,
-            newres: newres,
-            res: res,
-            locale: "ko-KR",
-            dedupByBaseTranslation: true,
-            deviceType: undefined,
-            API: makeApi(),
-            makeLookupParams: makeLookupParams
-        }, function() {
-            expect(resFileType._files["ko-KR"]).toBeUndefined();
-            expect(newres.add).toHaveBeenCalledTimes(1);
-            done();
-        });
-    });
-
-    // ── direct hit paths ──────────────────────────────────────────────────────
-
-    test("direct hit + source mismatch → addNewResource", function(done) {
-        var res = makeTestResource("hello", "Hello Updated");
-        var translated = makeTranslated("안녕", "Hello Old", "Hello Old");
-        var db = {
-            getResourceByCleanHashKey: jest.fn(function(key, cb) {
-                cb(null, key === "hello::ko-KR" ? translated : null);
-            })
-        };
-        var resFileType = makeCollector();
-        var newres = makeNewres();
-        var makeLookupParams = function(resource, locale) {
-            return { db: db, resource: resource, locale: locale, policy: [] };
-        };
-
-        writeTranslatedResource({
-            db: db,
-            resFileType: resFileType,
-            newres: newres,
-            res: res,
-            locale: "ko-KR",
-            dedupByBaseTranslation: false,
-            deviceType: undefined,
-            API: makeApi(),
-            makeLookupParams: makeLookupParams
-        }, function() {
-            expect(resFileType._files["ko-KR"]).toBeUndefined();
-            expect(newres.add).toHaveBeenCalledTimes(1);
-            done();
-        });
-    });
-
-    test("direct hit + reskey mismatch → clone with updated reskey + addResource", function(done) {
-        var res = makeTestResource("hello", "Hello");
-        var translated = makeTranslated("안녕", "Hello", "Hello");
-        // translated.reskey differs from res.reskey to trigger clone path
-        translated.reskey = "tr-Hello";
-        var db = {
-            getResourceByCleanHashKey: jest.fn(function(key, cb) {
-                cb(null, key === "hello::ko-KR" ? translated : null);
-            })
-        };
-        var resFileType = makeCollector();
-        var newres = makeNewres();
-        var makeLookupParams = function(resource, locale) {
-            return { db: db, resource: resource, locale: locale, policy: [] };
-        };
-
-        writeTranslatedResource({
-            db: db,
-            resFileType: resFileType,
-            newres: newres,
-            res: res,
-            locale: "ko-KR",
-            dedupByBaseTranslation: false,
-            deviceType: undefined,
-            API: makeApi(),
-            makeLookupParams: makeLookupParams
-        }, function() {
-            expect(resFileType._files["ko-KR"].resources.length).toBe(1);
-            expect(resFileType._files["ko-KR"].resources[0].reskey).toBe(res.reskey);
-            expect(newres.add).not.toHaveBeenCalled();
-            done();
-        });
-    });
-});
