+++ packages/ilib-loctool-webos-cpp/CppFileType.js
@@ -22,7 +22,7 @@ var path = require("path");
 var CppFile = require("./CppFile.js");
 var JsonResourceFileType = require("ilib-loctool-webos-json-resource");
 var Utils = require("loctool/lib/utils.js")
-var { utils: pluginUtils, lookupByPolicy: lookupUtils } = require("ilib-loctool-webos-common");
+var { utils: pluginUtils, translationResolver: { buildResolver, resolveTranslation } } = require("ilib-loctool-webos-common");
 
 var CppFileType = function(project) {
     this.type = "cpp";
@@ -121,10 +121,7 @@ CppFileType.prototype.write = function(translations, locales) {
             return locale !== this.project.sourceLocale && locale !== this.project.pseudoLocale;
         }.bind(this));
 
-    lookupUtils.detectCommonData(this, translations);
-
-    var policy = lookupUtils.buildPolicy();
-    var makeLookupParams = lookupUtils.createLookupParams(this, db, policy);
+    var resolver = buildResolver(db, translations);
 
     if (mode === "localize") {
         for (var i = 0; i < resources.length; i++) {
@@ -135,8 +132,8 @@ CppFileType.prototype.write = function(translations, locales) {
                 this.logger.trace("Localizing C++ strings to " + locale);
 
                 customInheritLocale = this.project.getLocaleInherit(locale);
-                lookupUtils.writeTranslatedResource({
-                    db: db,
+                resolveTranslation({
+                    resolver: resolver,
                     resFileType: resFileType,
                     newres: this.newres,
                     res: res,
@@ -145,8 +142,7 @@ CppFileType.prototype.write = function(translations, locales) {
                     dedupByBaseTranslation: true,
                     translationLocales: translationLocales,
                     deviceType: deviceType,
-                    API: this.API,
-                    makeLookupParams: makeLookupParams
+                    API: this.API
                 });
             }.bind(this));
         }
